Rm step for installing `musl-tool` in job `test`
The job `test` for target `x86_64-unknown-linux-musl` requires
`musl-tool` since there is no pre-built binary present for that target
and thus `cargo-binstall` automatically fall back to `cargo-build`.

Currently, the test for it also failed because `libssl-1.1` is not
present.

However, since this will be fixed once a new release of `cargo-binstall`
is published, this step should not be present.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>