Remove support for .crates2.json
The extra information it contains and requires is nonsensical for Binstall.
Instead of faking the details, which could potentially be harmful to cargo-install,
we will drop support for v2 and only write to v1.

Later we will add support for our own manifest and lockfile format that carries
relevant information for Binstall and fills some other gaps these manifests do not,
like local installs. See #176.