Optimize `Template::parse_items`
Replace the algorithm with explicit state and complex matching with a
loop that encodes state in the current structure of the loop.

The new code is much more readable and might be more performant since it
uses `str::find`/`str::split_once` to look for next token instead of
manually iterating over `str::chars`.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>