Fix `leon::ValuesFn`: impl for `Fn(&str) -> Option<Cow<'static, str>>`
If we impl it for `<'f> Fn(&str) -> Option<Cow<'f, str>> + 'f`, then it
would somehow imply it to `'static` when using it in `benches/values.rs`
and `benches/others.rs`, thus I decided to simplify it back to only
implement it for `'static` string.

While less flexible than `<'s> Fn(&'s str) -> Option<Cow<'s, str>>`,
this is enough for most use cases and produce a much simpler `Values`
trait.

Users who wants more flexibility should implement `Values` themselves.

This commit also extracts `benches` as a separate crate in an
independent workspace to avoid building criterion and tinytemplate in
CI, which makes it much slower as more crates need to be built and
criterion actually pulls in clap, and a whole lots of other crates.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>